+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Fix bug in template of path when decoding direct array items
+
 ## v0.7.0 (April 30, 2024)
 
 - Add optional `transform` argument to `decode` function